+++ internal/attributes/age/age.go
@@ -29,11 +29,6 @@ func NewAgeStore(c Config, sourceDir string) (*AgeStore, error) {
 		return nil, err
 	}
 	var a AgeStore
-	dir := filepath.Dir(c.IdentPath)
-	// TODO this was added for testing, is it needed?
-	if dir == "." {
-		c.IdentPath = filepath.Join(sourceDir, c.IdentPath)
-	}
 	ifile, err := os.Open(c.IdentPath)
 	if err != nil {
 		return nil, err